feat: Add coverage badge to CI and README
Added a new CI step to generate a coverage report and a badge. The coverage
badge is then committed to the repository if it's on the main branch. This
change also updates the README to include the coverage badge. A new `coverage`
target was added to the Makefile to facilitate this.

Signed-off-by: Chmouel Boudjnah <chmouel@chmouel.com>